Fix compile errors from cmake move
Reviewed-on: http://review.couchbase.org/30476
Tested-by: Michael Wiederhold <mike@couchbase.com>

+++ src/common.h
@@ -164,11 +164,11 @@ inline bool parseUint64(const char *str, uint64_t *out) {
     errno = 0;
     *out = 0;
     char *endptr;
-    unsigned long long ull = strtoull(str, &endptr, 10);
+    uint64_t ull = strtoull(str, &endptr, 10);
     if (errno == ERANGE)
         return false;
     if (xisspace(*endptr) || (*endptr == '\0' && endptr != str)) {
-        if ((long long) ull < 0) {
+        if ((int64_t) ull < 0) {
             /* only check for negative signs in the uncommon case when
              * the unsigned number is so big that it's negative as a
              * signed number. */
